150 back implemeting private channels

+++ api/src/channel/channel.service.ts
@@ -7,9 +7,26 @@ import { UserService } from 'src/user/user.service';
 import { getRepository, Repository } from 'typeorm';
 import { Channel } from './channel.entity';
 import { validate as isValidUUID } from 'uuid';
-import { Request } from 'express';
+import { Request, Response } from 'express';
 import { ModifyChannelDto } from 'src/dtos/modifyChannel.dto';
 import * as bcrypt from 'bcrypt';
+import { ExceptionFilter } from '@nestjs/common';
+import { ArgumentsHost } from '@nestjs/common';
+import { Catch } from '@nestjs/common';
+import { QueryFailedError } from 'typeorm';
+
+
+@Catch(HttpException)
+export class HttpExceptionsFilter implements ExceptionFilter {
+  catch(exception: any, host: ArgumentsHost) {
+	  console.log(exception)
+		console.log ( "coucou laaa ")
+	  const test : Response = host.switchToHttp().getResponse<Response>();	  //const response: Response = ctx.getResponse<Response>();
+	  test.status(100);
+	 // console.log("REEPONSE IS " + response + "haha")
+	//return res.status("LOL");
+  }
+}
 
 
 @Injectable()
@@ -25,17 +42,18 @@ export class ChannelService {
 	 * @param req the request containing user id
 	 * @returns
 	 */
-	public async createChannel(name: string, user : User, password: string = null)
+	public async createChannel(name: string, user: User, password: string = null, privacy : boolean = false)
 	{
 		const chan: Channel = new Channel();
 		chan.name = name;
 		chan.owner = user;
+		chan.private = privacy;
 		if (password)
 		{
 			chan.password_protected = true;
 			chan.password = await bcrypt.hash(password, 10);
 		}
-		await this.channelsRepo.save(chan);
+		await this.channelsRepo.save(chan)
 		return await this.userService.joinChannel(user, name, password);
 	}
 
@@ -50,7 +68,23 @@ export class ChannelService {
 			.leftJoinAndSelect("Channel.users", "Users")
 			.leftJoinAndSelect("Channel.banned", "b")
 			.leftJoinAndSelect("Channel.owner", "o")
-		.getMany();
+			.getMany();
+	}
+
+	/**
+	 * @brief Returns all users of all existing channels
+	 * @returns
+	 */
+	public async getChannelsForUser(user : User) : Promise<Channel[]>
+	{
+		return await this.channelsRepo.createQueryBuilder('channel')
+			.orderBy("channel.name")
+			.leftJoinAndSelect("channel.users", "Users")
+			.leftJoinAndSelect("channel.banned", "b")
+			.leftJoinAndSelect("channel.owner", "o")
+			.where('channel.private = false')
+			.orWhere("Users.id = :id", { id: user.id })
+			.getMany();
 	}
 
 	/**
@@ -62,18 +96,17 @@ export class ChannelService {
 	{
 		let chan : Channel = await this.channelsRepo.findOne({ where: { name: channelIdentifier }, relations: ['messages', 'banned'] });
 		if (!chan && isValidUUID(channelIdentifier))
-			await this.channelsRepo.findOne({ where: { id: channelIdentifier }, relations: ['messages', 'banned'] });
+			await this.channelsRepo.findOne({ where: { id: channelIdentifier }, relations: ['messages', 'banned'] })
 		if (!chan)
-			throw new HttpException('Channel not found (id or name)', HttpStatus.NOT_FOUND);
+			throw new HttpException('Channel ' + channelIdentifier + ' not found (id or name)', HttpStatus.NOT_FOUND);
 		return chan;
 	}
 
 	async	getChannelPasswordHash(channelId: string): Promise<string> {
 		const chan: Channel = await this.channelsRepo.createQueryBuilder('Channel')
-		.select(["Channel.password"])
-		.where({ "id": channelId})
-		.getOne();
-
+			.select(["Channel.password"])
+			.where({ "id": channelId })
+			.getOne();
 		return chan.password;
 	}
 
@@ -98,15 +131,17 @@ export class ChannelService {
  	* @param changes changes to be performed - chanName or ownership
  	* @returns All channels after deletion
  	*/
-	 public async deleteChannel(user: User, channel: Channel) : Promise<Channel[]>
+
+	 public async deleteChannel(user: User, channel: Channel)// : Promise<Channel[]>
 	 {
+		if (channel.ownerId != user.id)
+			throw new HttpException("You must be admin to delete chan.", HttpStatus.FORBIDDEN);
 		await this.channelsRepo
     		.createQueryBuilder()
     		.delete()
     		.from(Channel)
     		.where("name = :channame", { channame: channel.name })
 			.execute();
-		 return await this.getUsersOfChannels();
 	}
 
 
@@ -118,6 +153,8 @@ export class ChannelService {
  	*/
 	public async deleteUserFromChannel(user: User, channel : Channel, toBan: User) : Promise<Channel[]>
 	{
+		if (channel.ownerId != user.id)
+			throw new HttpException("You must be admin to delete chan.", HttpStatus.FORBIDDEN);
 		await this.channelsRepo.createQueryBuilder()
 			.relation(Channel, "users")
 			.of({ id: toBan.id })
@@ -136,6 +173,8 @@ export class ChannelService {
 
 	public async temporaryBanUser(user: User, channel: Channel, toBan: User)
 	{
+		if (channel.ownerId != user.id)
+			throw new HttpException("You must be admin to delete chan.", HttpStatus.FORBIDDEN);
 		console.log("Bannishement");
 		/** Step one : Deleting user from channel */
 		await this.deleteUserFromChannel(user, channel, toBan);
